+++ src/extension.ts
@@ -19,18 +19,21 @@ import { HolidayHelper } from './shared/holidayHelper';
 import { LeekFundConfig } from './shared/leekConfig';
 import { Telemetry } from './shared/telemetry';
 import { SortType } from './shared/typed';
-import { formatDate, isStockTime } from './shared/utils';
+import { events, formatDate, isStockTime } from './shared/utils';
 import { StatusBar } from './statusbar/statusBar';
 import { ProfitStatusBar } from './statusbar/Profit';
 import { cacheStocksRemindData } from './webview/leekCenterView';
 import { cacheFundAmountData, updateAmount } from './webview/setAmount';
-import { events } from './shared/utils';
 import FlashNewsOutputServer from './output/flash-news/FlashNewsOutputServer';
+import { ForexService } from './explorer/forexService';
+import { ForexProvider } from './explorer/forexProvider';
 
 let loopTimer: NodeJS.Timer | null = null;
 let binanceLoopTimer: NodeJS.Timer | null = null;
+let forexLoopTimer: NodeJS.Timer | null = null;
 let fundTreeView: TreeView<any> | null = null;
 let stockTreeView: TreeView<any> | null = null;
+let forexTreeView: TreeView<any> | null = null;
 let binanceTreeView: TreeView<any> | null = null;
 
 let flashNewsOutputServer: FlashNewsOutputServer | null = null;
@@ -61,10 +64,12 @@ export function activate(context: ExtensionContext) {
   const fundService = new FundService(context);
   const stockService = new StockService(context);
   const binanceService = new BinanceService(context);
+  const forexService = new ForexService(context);
 
   const nodeFundProvider = new FundProvider(fundService);
   const nodeStockProvider = new StockProvider(stockService);
   const binanceProvider = new BinanceProvider(binanceService);
+  const forexProvider = new ForexProvider(forexService);
   const newsProvider = new NewsProvider();
 
   const statusBar = new StatusBar(stockService, fundService);
@@ -81,6 +86,10 @@ export function activate(context: ExtensionContext) {
 
   binanceTreeView = window.createTreeView('leekFundView.binance', {
     treeDataProvider: binanceProvider,
+  });
+
+  forexTreeView = window.createTreeView('leekFundView.forex', {
+    treeDataProvider: forexProvider,
   });
 
   window.createTreeView('leekFundView.news', {
@@ -108,6 +117,7 @@ export function activate(context: ExtensionContext) {
         setIntervalTime();
         return;
       }
+
       if (fundTreeView?.visible) {
         // fix https://github.com/giscafer/leek-fund/issues/78
         if (globalState.fundAmountCacheDate !== formatDate(new Date())) {
@@ -157,6 +167,17 @@ export function activate(context: ExtensionContext) {
       // intervalTimeConfig < 3000 ? 3000 : intervalTimeConfig
       300000 // 该功能存在网络问题（一些网络有vpn都无法请求通），这里故意设置长时间
     );
+
+    /* 汇率变化轮询间隔2分钟 */
+    if (forexLoopTimer) {
+      clearTimeout(forexLoopTimer);
+      forexLoopTimer = null;
+    }
+    forexLoopTimer = setInterval(() => {
+      if (forexTreeView?.visible) {
+        forexProvider.refresh();
+      }
+    }, 120000);
   };
 
   setIntervalTime();
@@ -171,6 +192,7 @@ export function activate(context: ExtensionContext) {
     nodeStockProvider.refresh();
     newsProvider.refresh();
     binanceProvider.refresh();
+    forexProvider.refresh();
     flashNewsOutputServer?.reload();
     events.emit('onDidChangeConfiguration');
     profitBar?.reload();
@@ -185,7 +207,8 @@ export function activate(context: ExtensionContext) {
     nodeStockProvider,
     newsProvider,
     flashNewsOutputServer,
-    binanceProvider
+    binanceProvider,
+    forexProvider
   );
 
   // Telemetry Event
@@ -193,8 +216,7 @@ export function activate(context: ExtensionContext) {
 }
 
 function setGlobalVariable() {
-  const iconType = LeekFundConfig.getConfig('leek-fund.iconType') || 'arrow';
-  globalState.iconType = iconType;
+  globalState.iconType = LeekFundConfig.getConfig('leek-fund.iconType') || 'arrow';
 
   const fundAmount = LeekFundConfig.getConfig('leek-fund.fundAmount') || {};
   cacheFundAmountData(fundAmount);
@@ -202,18 +224,15 @@ function setGlobalVariable() {
   const stocksRemind = LeekFundConfig.getConfig('leek-fund.stocksRemind') || {};
   cacheStocksRemindData(stocksRemind);
 
-  const showEarnings = LeekFundConfig.getConfig('leek-fund.showEarnings');
-  globalState.showEarnings = showEarnings;
+  globalState.showEarnings = LeekFundConfig.getConfig('leek-fund.showEarnings');
 
-  const remindSwitch = LeekFundConfig.getConfig('leek-fund.stockRemindSwitch');
-  globalState.remindSwitch = remindSwitch;
+  globalState.remindSwitch = LeekFundConfig.getConfig('leek-fund.stockRemindSwitch');
 
   globalState.labelFormat = LeekFundConfig.getConfig('leek-fund.labelFormat');
 
   globalState.immersiveBackground = LeekFundConfig.getConfig('leek-fund.immersiveBackground', true);
 
-  const fundGroups = LeekFundConfig.getConfig('leek-fund.fundGroups') || [];
-  globalState.fundGroups = fundGroups;
+  globalState.fundGroups = LeekFundConfig.getConfig('leek-fund.fundGroups') || [];
 
   const fundLists = LeekFundConfig.getConfig('leek-fund.funds') || [];
   if (typeof fundLists[0] === 'string' || fundLists[0] instanceof String) {