+++ src/registerCommand.ts
@@ -2,6 +2,7 @@ import { commands, ExtensionContext, window } from 'vscode';
 import fundSuggestList from './data/fundSuggestData';
 import { BinanceProvider } from './explorer/binanceProvider';
 import BinanceService from './explorer/binanceService';
+import { ForexProvider } from './explorer/forexProvider';
 import { FundProvider } from './explorer/fundProvider';
 import FundService from './explorer/fundService';
 import { NewsProvider } from './explorer/newsProvider';
@@ -36,7 +37,8 @@ export function registerViewEvent(
   stockProvider: StockProvider,
   newsProvider: NewsProvider,
   flashNewsOutputServer: FlashNewsOutputServer,
-  binanceProvider?: BinanceProvider
+  binanceProvider: BinanceProvider,
+  forexProvider: ForexProvider
 ) {
   const newsService = new NewsService();
   const binanceService = new BinanceService(context);
@@ -44,7 +46,7 @@ export function registerViewEvent(
   commands.registerCommand('leek-fund.toggleFlashNews', () => {
     const isEnable = LeekFundConfig.getConfig('leek-fund.flash-news');
     LeekFundConfig.setConfig('leek-fund.flash-news', !isEnable).then(() => {
-      window.showInformationMessage(`已${isEnable ? '启用' : '关闭'} OUTPUT 的 Flash News！`);
+      window.showInformationMessage(`已${isEnable ? '关闭' : '启用'} OUTPUT 的 Flash News！`);
     });
   });
 
@@ -179,9 +181,7 @@ export function registerViewEvent(
         return;
       }
       // 存储到配置的时候是接口的参数格式，接口请求时不需要再转换
-      const newCode = code
-        .replace('gb', 'gb_')
-        .replace('us', 'usr_');
+      const newCode = code.replace('gb', 'gb_').replace('us', 'usr_');
       LeekFundConfig.updateStockCfg(newCode, () => {
         stockProvider.refresh();
       });
@@ -347,6 +347,14 @@ export function registerViewEvent(
       binanceProvider?.refresh();
     });
   });
+
+  /**
+   * Forex command
+   */
+  commands.registerCommand('leek-fund.refreshForex', () => {
+    forexProvider.refresh();
+  });
+
   /**
    * Settings command
    */